// src/app/enums/ELanguage.ts
export enum Languages {
  JavasScript = 'JavaScript',
  TypeScript = 'TypeScript',
  Java = 'Java',
  CSharp = 'C#',
  CPlusPlus = 'C++',
  TailwindCSS = 'TailwindCSS',
  Firebase = 'Firebase',
  Bootstrap = 'Bootstrap',
  CSS = 'CSS',
  HTML = 'HTML',
  SWIFT = 'SWIFT',
  Angular = 'Angular',
  Python = 'Python',
  Sklearn = 'Sklearn',
  Pandas = 'Pandas',
  Numpy = 'Numpy',
  Matplotlib = 'Matplotlib',
  OpenCV = 'OpenCV',
  ThreeJS = 'ThreeJS',
  TenserFlow = 'TenserFlow',
  Keras = 'Keras',
  AndroidStudio = 'Android Studio',
  Kotlin = 'Kotlin',
  API = 'API',
  C = 'C',
  CouchDB = 'CouchDB',
  Ionic = 'Ionic',
  Git = 'Git',
  ReactNative = 'React Native',
  Paper = 'Paper',
  Expo = 'Expo',
  Quarkus = 'Quarkus',
  MangoDB = 'MangoDB',
  RxJS = 'RxJS',
  Cypress = 'Cypress',
  Laravel = 'Laravel',
  PHP = 'PHP',
  AFrame = 'A-Frame',
  SCRUM = 'SCRUM',
  Net = '.NET',
  Genexus = 'Genexus',
  PostgreSQL = 'PostgreSQL',
  Oracle = 'Oracle',
  MicrosoftSQLServer = 'Microsoft SQL Server',
  Html = 'Html 5',
  Css = 'Css ',
  Customer = 'Customer service and support',
  Hardware = 'Hardware maintenance',
  SpringBoot = 'Spring Boot',
  React = 'React JS',
  Vite = 'Vite'
}